Refactor datetime aggregation methods and remove redundant code in Analysis class
